+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- `pageInfo` event for custom search pages.
+
 ## [3.122.6] - 2023-06-26
 
 ### Changed